Ditch docker for the frontend (more trouble than it’s worth)
We already have a proxy to the locally running backend set up with Vite
— the frontend talks to the containerized backend no problem.

What _is_ a problem is that the container and host have different
architectures, and we want `node_modules` on both.  On the host for
developer experience, and inside the container for serving the app.
Better we just serve the app directly from the host and skip the
trouble.